pyproject.toml
[project]
name = "dbd-player-insights"
version = "0.1.0"
description = "Dead by Daylight telemetry data pipeline — POC for Behaviour Interactive"
requires-python = ">=3.11"
dependencies = [
    "orjson>=3.9",
    "faker>=22.0",
    "boto3>=1.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]
dbt = [
    "dbt-core>=1.7",
    "dbt-snowflake>=1.7",
]
dashboard = [
    "streamlit>=1.30",
    "plotly>=5.18",
    "snowflake-connector-python>=3.6",
]

[project.scripts]
dbd-generate = "generator.cli:main"

[tool.setuptools.packages.find]
include = ["generator*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["generator/tests"]
